+++ src/components/ui-form.ts
@@ -4,7 +4,7 @@
 // @copyright   : 2016 Sigma Frameworks
 // @license     : MIT
 
-import {autoinject, customElement, bindable, inlineView, bindingMode} from "aurelia-framework";
+import {autoinject, customElement, bindable, inlineView, bindingMode, TaskQueue} from "aurelia-framework";
 import {UIEvent} from "../utils/ui-event";
 import {_, UIUtils} from "../utils/ui-utils";
 
@@ -16,17 +16,16 @@ export class UIForm {
 
     private __form: HTMLElement;
 
-    constructor(public element: Element) {
+    constructor(public element: Element, public taskQueue: TaskQueue) {
         if (!this.element.hasAttribute('auto-grid')) this.element.classList.add('two-column');
     }
 
     attached() {
-        setTimeout(() => {
+        this.taskQueue.queueMicroTask(() => {
             let el: any = this.element.querySelector('ui-input input,textarea,ui-phone input');
             if (!isEmpty(el)) el.focus();
-        }, 10);
-
-        if (this.busy) setTimeout(() => this.busyChanged(true), 200);
+            if (this.busy) this.busyChanged(true);
+        });
     }
 
     busyChanged(newValue: any) {