+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added support to use em/rem values for Dialog width/height and Data-Column widths
+* Updated validation-renderer
 
 <a name="0.0.1-beta.1"></a>
 ## 0.0.1-beta.1 (2016-08-23)